+++ src/Summoner/CLI.hs
@@ -25,7 +25,8 @@ import Summoner.Config (ConfigP (..), PartialConfig, defaultConfig, finalise, lo
 import Summoner.Decision (Decision (..))
 import Summoner.Default (defaultConfigFile)
 import Summoner.GhcVer (GhcVer, showGhcVer)
-import Summoner.License (License (..), LicenseName (..), fetchLicense, parseLicenseName)
+import Summoner.License (License (..), LicenseName (..), fetchLicense, licenseShortDesc,
+                         parseLicenseName)
 import Summoner.Project (generateProject)
 import Summoner.Settings (CustomPrelude (..))
 import Summoner.Validation (Validation (..))
@@ -50,7 +51,7 @@ runShow = \case
         -- show list of all available GHC versions
         GhcList -> showBulletList @GhcVer showGhcVer (reverse universe)
         -- show a list of all available licenses
-        LicenseList Nothing -> showBulletList @LicenseName show universe
+        LicenseList Nothing -> showBulletList @LicenseName showDesc universe
         -- show a specific license
         LicenseList (Just name) ->
             case parseLicenseName (toText name) of
@@ -65,6 +66,9 @@ runShow = \case
   where
     showBulletList :: (a -> Text) -> [a] -> IO ()
     showBulletList showT = mapM_ (infoMessage . T.append "➤ " . showT)
+
+    showDesc :: LicenseName -> Text
+    showDesc l = show l <> ": " <> licenseShortDesc l
 
 runNew :: NewOpts -> IO ()
 runNew NewOpts{..} = do